requirements.txt
astroid==2.2.5
dj-database-url==0.5.0
Django==2.1.8
django-filter==2.1.0
django-heroku==0.3.1
djangorestframework==3.9.4
ez-setup==0.9
graphviz==0.11
gunicorn==19.9.0
isort==4.3.20
lazy-object-proxy==0.0.0
mccabe==0.6.1
psycopg2==2.8.2
pylint==2.3.1
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
whitenoise==4.1.2
wrapt==1.11.1
